requirements.txt
pygame
PyYAML
opencv-python
mediapipe
numpy
setuptools
wheel
